requirements.txt
Flask
python-dotenv
langchain-community
google-generativeai
pypdf
sentence-transformers
faiss-cpu